+++ CHANGELOG.md
@@ -37,6 +37,7 @@ In addition to original Keep-a-Changelog, we use following rules:
 
 ### Changed
 - Implement `AsRef` and `AsMut` for `XXAny`. https://github.com/ricosjp/ruststep/pull/180
+- Cut out `IntoOwned` from `Holder` for compile `LIST OF LIST OF XXX`. https://github.com/ricosjp/ruststep/pull/183
 - Implement `AsRef` and `AsMut` in the case of multiple inheritance. https://github.com/ricosjp/ruststep/pull/179
 - Implements `AsRef`, `AsMut`, `Deref`, `DerefMut` for subtypes. https://github.com/ricosjp/ruststep/pull/174
 - Implements `Serialize` to `ruststep::primitive::Logical`. https://github.com/ricosjp/ruststep/pull/170